Fixing the tests
