+++ client/client_test.go
@@ -1,43 +1,12 @@
 package client
 
 import (
-	"fmt"
-	"net/http"
-	"net/url"
-	"testing"
-
-	"github.com/hashicorp/go-oracle-terraform/opc"
-	"github.com/kylelemons/godebug/pretty"
 	"gopkg.in/jarcoal/httpmock.v1"
+	"net/url"
+	"net/http"
+	"testing"
+	"github.com/hashicorp/go-oracle-terraform/opc"
 )
-
-func TestClient_qualifyList(t *testing.T) {
-	client, server, err := getBlankTestClient()
-	if err != nil {
-		t.Fatal(err)
-	}
-	defer server.Close()
-
-	input := []string{
-		"foo",
-		"bar",
-		"baz",
-	}
-
-	baseStr := fmt.Sprintf("/Compute-%s/%s", _ClientTestDomain, _ClientTestUser)
-
-	expected := []string{
-		fmt.Sprintf("%s/%s", baseStr, "foo"),
-		fmt.Sprintf("%s/%s", baseStr, "bar"),
-		fmt.Sprintf("%s/%s", baseStr, "baz"),
-	}
-
-	result := client.getQualifiedList(input)
-
-	if diff := pretty.Compare(result, expected); diff != "" {
-		t.Fatalf("Qualified List Diff: (-got +want)\n%s", diff)
-	}
-}
 
 func TestClient_retryHTTP(t *testing.T) {
 	httpmock.Activate()
@@ -49,10 +18,10 @@ func TestClient_retryHTTP(t *testing.T) {
 	}
 
 	client := Client{}
-	client.maxRetries = opc.Int(5)
+	client.MaxRetries = opc.Int(5)
 	// Can't use a custom transport, otherwise httpmock won't catch request
 	client.httpClient = http.DefaultClient
-	client.apiEndpoint = endpoint
+	client.APIEndpoint = endpoint
 	client.logger = opc.NewDefaultLogger()
 	client.loglevel = opc.LogLevel()
